ideas on how to refactor
